pyproject.toml
[tool.poetry]
name = "geoagency"
version = "0.1.0"
description = ""
authors = ["simeonwetzel <simeon.wetzel@tu-dresden.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
smolagents = "^1.4.1"
fastapi = {extras = ["standard"], version = "^0.115.6"}
pydantic = "^2.10.5"
loguru = "^0.7.3"
litellm = "^1.59.1"
aiohttp = "^3.11.11"
nest-asyncio = "^1.6.0"
flashrank = "^0.2.10"
shapely = "^2.0.7"
tomark = "^0.1.4"
gradio = "^5.15.0"
json5 = "^0.10.0"
arize-phoenix = "^8.27.0"
opentelemetry-sdk = "^1.32.1"
opentelemetry-exporter-otlp = "^1.32.1"
openinference-instrumentation-smolagents = ">=0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
